Fix Throttler initialization.
Ensure that the Throttler last allowed time is zero-initialized. This
field was uninitialized on macOS with clang 14, causing test failures.

Since we are here, we also switch the Throttler to use the chrono
steady_clock (since we don't want it to jump during a throttle interval),
and remove virtual functino annotations, since Throttler can't reasonably
serve as a base clase.

Signed-off-by: James Peach <jpeach@apache.org>