fix KeyEqual param for ConnectingIpPool declaration (#10157)
The 4th parameter, KeyEqual, is missing in ConnectingIpPool declaration, which may lead to the comparison of two IpEndpoint object won't work as we expected. That is, the default one is unlikely to work with sockaddr type.
In another way, IpHelper contains the compare operator for IpEndpoint type, so IpHelper should be used as the KeyEqual parameter of ConnectingIpPool declaration.