Remove Http3NoError allocations
The Http3Error mechanism is written around dynamic polymorphism, with
Http3Error as the base class and a few child classes for specifc types
of errors. One of those was Http3NoError. It was observed that the
frequent allocations for Http3NoError in many places was wasteful, so
this patch removes those allocations by instead returning nullptr
instead and removes the Http3NoError class since it is no longer used.

Fixes #9992