cache_range test performance improvement (#10170)
cache_range_requests_cache_complete_responses.test.py had a for loop
where it appended a single character to a string 4 million times. This
is exceedingly slow in Python as it needed to reallocate, copy, and
append with each iteration. This patched version reduces the time to a
fraction of a second instead of multiple minutes (4 minutes on my
system).