Add version range up to CMake 3.27 (#10259)
Specifying a version range ensures that our build is compatible with NEW
policies in future CMake versions, and therefore won't unexpectedly
break when we increase our mininum version. It is fine to use a CMake
version newer than the upper end of our version range, but we should
keep the range updated as newer versions become available. We should use
a latest CMake on at least one CI run.