Remove TSDebug() and TSIsDebugTagSet() TS API functions.
Replace with Dbg() and DbgCtl::on().  This required
extensive changes to the ESI plugin unit tests.

Doc changes were also necessary.

The redirect_1 example plugin is changed to illustrate
how to implement debug trace output in a plugin that
can compile and run with either these ATS10 changes
or for older versions of ATS.