Import of the Cripts scripting / plugin wrapper library and includes (#10191)
* Import of the Cripts scripting / plugin wrapper library and includes

Cripts tries to solves 90% of your problems, making it simple to
do the most common tasks. It achieves this by managing the data
needed for such tasks, using a per transaction context.

It doesn't support complex operations such as transforms, or server
intercepts. It also only supports building with CMake in this version.

* Prevent RAT from complaining on clang-tidy for Cripts

* Make Cripts optional, default off