Traffic Dump: fix a use after free of a mutex
Coverity found that Traffic Dump had code that locked a mutex via
std::lock_guard, freed the owner of the mutex (and thus the mutex), and
then, when the std::lock_guard went out of scope, unlocked the free'd
mutex. This use after free is addressed in this patch by extending the
lifetime of the SessionData object owning the mutex until after the lock
is released.

Fixes: #10331