Build unit tests only with BUILD_TESTING (#10548)
* Build unit tests only with BUILD_TESTING

This makes all the tests conditional on BUILD_TESTING being set.
Respecting BUILD_TESTING is a good idea because it is officially
documented as part of CTest and people will expect it to work. Also,
until now we have no way to not build the tests if we should want to.

* Remove enable_testing() calls

The command is automatically invoked when we include CTest.