Add ENABLE_CXX_20 option
There are multiple ways to approach this. Since the ability to build
with C++20 is intended for CI, I opted to add an option that can be
preset. An alternative that I tested was to simply respect
CMAKE_CXX_STANDARD if it was set, but I don't like this, because it
doesn't work on most projects and consistency is good. I also think the
option makes the purpose explicit, whereas allowing CMAKE_CXX_STANDARD
to be set from the command line is more flexible, but more mysterious.

Fixes #10609