Break cycle between iocore/inknet and proxy/http (#10587)
Summary of changes:

* Remove unused headers.
* Break the PreWarm namespace out of PreWarmManager and leave it in iocore/net.
* Move PreWarmManager back to proxy/http (I had moved it to iocore earlier).
* Use a callback to reconfigure the prewarm manager when sni.yaml is reloaded.

Removing the unused headers is simply cleanup. They have been in the source code for a long time and aren't currently used for anything. The other two changes were more significant. I think there's a tradeoff between putting the PreWarm namespace in iocore/net or proxy/http. The former makes it easy to avoid the dependency cycle, but I think the code is more closely related to the rest of the PreWarmManager than to iocore/net. The last change was to address the issue of reloading the pre-warm configuration when the SNI was reloaded, because SNI rules could change pre-warm config. The codebase currently handled this by calling reconfigure() on the global prewarmManager, so I proposed to @masaori335 that the prewarmManager could instead register its reconfigure() method as a callback. Because that method is static, I think this ended up being kind of messy.