Extract CacheEvacuateDocVC from CacheVC (#10603)
This extracts a subclass from CacheVC called CacheEvacuateDocVC and moves the logic for evacuating a doc to the new subclass. By having less code in it, CacheVC should be a little simpler. While doing the extraction, I also read through the logic and tried to clean up the logic flow a little bit. This last bit was risky, and I'm not confident we have enough test coverage to be sure it was done correctly. I did my best to double check my reasoning on every change, but please review carefully.