Improve the TSMgmtUpdateRegister with an optional file name
Providing this filename makes the update registry keep track
of file changes automatically. When used, the continuation
provided will only be called if changes to the configuration
file is detected.

Also, this eliminates the superflous code for calling a named
plugin reload. That code is never used, and I don't like having
code there "just because", when nothing uses it.