Fix race of shutdown & retry of CacheDir unit test (#10659)
Fix #10657.

The cache_init_success_callback can be called twice when the try-lock is failed and retry event is scheduled. It can be cause of double free.

To fix this, we should delete CacheDirTest only if all test is done. Also, I thought we need TerminalTest continuation to shutdown the test, but it looks like we can just call test_done() function directly.