NetHandler: make additional_accepts reachable by accept threads
Accept threads may be on dedicated threads not managed by a NetHandler
instance. Prior to this change, the cases where a NetHandler was
accessed via such threads would have been undefined. To handle the
additional_accepts configuration, this makes that configuration atomic
and provides an accessor to it that can be accessed across all threads
that need the configured value.