Harden the tunnel_transform Autest (#10726)
The tunnel_transform test failed internally in certain environment as the actual TLS traffic byte counts fell out of the hard-coded range of expected value.

The TLS byte count can varied across different openssl/curl versions, so this PR adds a simple forwarding proxy to keep track of the exact byte count between the client and server, as suggested in the test comment. The exact byte count is used in the validation.

Note: originally I was trying a nc + pipe solution. It worked in my environment but I am concerned about the portability of different versions of nc and different characteristics of named pipe on different platforms(unidirectional VS bidirectional). So sticking with the hand-rolled proxy for portability.