Fix sni.yaml ip_allow tunnel_route (#10761)
When combining a rejection of an incoming connection via sni.yaml
ip_allow and tunnel_route, ATS still initiated an unneccessary
connection to the origin. This constituted an unused TLS handshake with
the origin because it would immediately be closed after the client side
handshake ended. This patch fixes ATS so it doesn't do anything on the
origin side if the client side TLS handshake failed already at the
CLIENT_HELLO.