Extract class AggregateWriteBuffer from Stripe (#10718)
This encapsulates the `agg_` fields of `Stripe` in a new class. This is
in preparation to move all the aggregation behavior into this new
class, including the behavior to enqueue virtual connections and
update the number of bytes waiting to be written to the buffer, which
is currently being done by the virtual connections themselves.

Note: `Stripe` does not use its own getters to access aggregate buffer,
but calls the getters on the `AggregateWriteBuffer` itself. This is
because I want to get rid of as many of the getters on `Stripe` as
possible.