cmake: Add -fPIC when checking for symbols
I noticed that on ubuntu:23:04 with ci-ubuntu, the symbol checks for
getresuid and the like were failing because the compilation of the
simple programs to check the symbols failed with complaints saying that
the code had to be compiled with PIC enabled. This caused cmake to
incorrectly think that these symbols didn't exist when in fact they did.
Adding -fPIC for the symbol checks fixed this.