stek_share: fix pthread_t print format warning (#10819)
For osx, pthread_t is a void* type rather than an int type. Thus on
these systems a warning was emitting about an inappropriate format
specifier for the pointer type argument. Talking with Fei, it looks like
we can simply remove that ID because it's not that helpful to have it in
the debug output.