Pull out down origin check from HttpSM::apply_ip_allow_filter
In a refactor on master for the ip_allow implementation in HttpSM
(see #9845) put into HttpSM::apply_ip_allow_filter. However, the down
origin feature is orthogonal to ip_allow and it is confusing to
implement its feature in a function that communicates that it implements
ip_allow filtering. This separates the down origin check from that
function like it was before the refactor.