yapf: use yapf instead of autopep8
autopep8 does not support Python 3.12, which is a requirement for moving
to fedora:39. The following ticket has been open without comment from
the development community for 6 weeks:

https://github.com/hhatto/autopep8/issues/712

Transitioning to yapf is easy, and the community behind it seems
stronger. I think moving forward it is a better Python formatter for us.

Conflicts:
      CMakeLists.txt
      tools/git/pre-commit