ja3_fingerprint: add modify-incoming option
This adds the optional modify-incoming argument to the
ja3_fingerprint.so plugin. This argument alters the header modification
behavior to add the X-JA3-* headers to the incoming client request
rather than to the outgoing proxy request. This allows other plugins
later in the plugin.config file to view the added headers as if they
were added by the client, which can be helpful to some plugins.