header_freq: Fix msg lock issues (#10917)
I recently tried to use header_freq but experienced crashes from failed
assertions due to the API hook logic not being able to take the lock for
the plugin's msg handler dispatch. This patch alleviates that by
removing the lock from the continuation since it contains no shared
continuation data and therefore doesn't need it. This should also make
the plugin less costly to run.

This also updates the docs to record how to log the stats to a specific
file since that was undocumented before. It also logs the headers such
that they are sorted.