multiplexer: set up transforms for PUT requests
Without this change, multiplexer handling of PUT requests only work for
very small bodies. This sets up the transform for PUT like we do for
POST. A future commit for large PUT/POST bodies will contain a test for
this. In the meantime, this at least connects up the transform for PUT
requests.