multiplexer: debug logging and variable name updates (#10964)
This is a multiplexer cleanup PR that adds debug logging to post.cc and
updates variable names for the transform in order to make their purpose
explicit (for the origin, for a copy, etc.). In post.cc, for instance,
both the incoming and outgoing VIO were named `vio`, with the difference
being whether it was called off of the state `s` variable or as a local
variable. This updates the names to clarify the purpose of each.

Other than logging changes, this patch makes no functional change to
multiplexer.