Transforms: Handle INT64_MAX TSVConnWrite values
Our example null_transform.cc and multiplexer plugins (and likely
others) perform their initial TSVConnWrite call with INT64_MAX, meaning
that nbytes should be set large at first and then adjusted later. It was
found with the multiplexer plugin that this works for small bodies, but
for larger bodies that spanned multiple WRITE_READY events, this
mechanism wasn't working. This was due to
setup_server_send_request_api() not being called when size was set to
INT64_MAX because it was enforcing a Content-Length header value. This
updates the HttpSM transform logic for these plugins so that it only
fails if the Content-Length header is actually missing from the server
side request.

This patch adds large body requests and responses to verify that
multiplexer handles them appropriately.