Add tscore to access_control dependencies (#10968)
This makes the ats_base64_decode symbol visible to the access_control
plugin. This symbol is not part of the API, so access_control should not
be using it, or it should be promoted to the API. This commit does not
solve that; it only fixes it to work like it did before.