multiplexer: Correctly Handle Large Content-Length POST/PUT Request Bodies (#10965)
The multiplexer plugin assumed that for POST/PUT requests it could pass INT64_MAX to the TSVConnWrite api for transforming the request bodies. After all, INT64_MAX is often used for transforming response bodies. INT64_MAX accidentally worked for small request bodies that got sent in a single READ_READY call, but for larger bodies, HttpSM complained about the unsupported INT64_MAX call for requests. It is unsupported for requests because INT64_MAX correlates with chunked requests which our transform mechanism doesn't support.

Thus, multiplexer was not correctly abiding by the expected API requirement that for POST/PUT requests, TSVConnWrite requires an explicit length passed in. This fixes that bug. We should consider for a future PR whether we want to support transforming chunked requests.