Observe cache_config_agg_write_backlog
The global `cache_config_agg_write_backlog` is defined as `2 * AGG_SIZE`.
There was an extra `+ AGG_SIZE` in `add_writer` with no comment; this
goes back all the way to the original first GitHub commit. This could be
a quick hack someone put in to fix an issue. If we need to keep this,
I'd like to add a comment explaining why, because it clearly violates
the intent of `cache_config_agg_write_backlog`. There is another place
in the code where `cache_config_agg_write_backlog` is used, in
`Stripe::open_write`, and in that place there is no extra `+ AGG_SIZE`.