Do away with temporary buffer
The temporary buffer may have been useful for performance years ago when
the allocation of the incoming and temporary buffers were slightly
different. Today, I can not see a reason why this temporary buffer would
be needed, and it seems to simplify the code a lot to do away with it.