Persist bad disks (#10960)
Add a config option to have traffic server remember which disks are bad. Previously, cache disks that were determined to be bad at runtime were forgotten upon restart. This could cause corrupt cache metadata to be loaded from disk, leading to crashes, or slowdowns when accessing the known bad disk.

If the option is turned on, bad disks are remembered until the known_bad_disks file is edited or deleted. If the system administrator replaces the disks, they need to perform this manual step. Because of this requirement, this option is off by default.