Move some logic from CacheVC to Stripe and AggregateWriteBuffer (#10987)
This breaks up some of handleRead into smaller methods and moves the logic for copying out of the aggregate write buffer to the classes responsible for that buffer.

There was a check to make sure that data was only copied from the aggregate write buffer if the requested data was available. I left the old check because it avoids allocating a new buffer if the data is unavailable, and I also duplicated it in Stripe::copy_from_aggregate_write_buffer for safety. I thought it would be better to have it in and take it out if it negatively impacts performance than not to have it and find out it would have been useful. I have not benchmarked this.

The order of the parameters of Stripe::copy_from_aggregate_write_buffer and AggregateWriteBuffer::copy_from was chosen to reflect the order of the memcpy parameters. This does not mean I like the way memcpy chose to do it, I just thought it might help consistency. :)

This is part of a chain of PRs to remove all the getters I added to Stripe, in order to encapsulate the aggregate write buffer properly. This PR removes Stripe::get_agg_buffer.

* Extract CacheVC::check_ram_cache from handleRead

* Extract check_last_open_read_call from handleRead

* Extract check_aggregation_buffer from handleRead

* Invert check_aggregation_buffer conditional logic

* Reorder buffer creation and agg_offset calculation

* Extract Stripe::copy_from_aggregate_write_buffer

* Extract AggregateWriteBuffer::copy_from

* Fix naming: check -> load_from

This fixes the names of the check_ CacheVC methods to reflect what they
actually do, which is to load from various caches into the CacheVC
buffer.

* Fix assertion: assert -> ink_assert

* Mark methods as const where appropriate