Revert "Coverity 1497378: Use of 32-bit time_t in CacheVC (#10857)"
On 64 bit systems, time_t (which is an unsigned long) is generally a 64
bit value. We therefore may not change the type of pin_in_cache from an
explicit 32 bit value to a 64 bit one without introducing a cache
inconsistency in ATS 10. Since the community is a hard pass on
introducing an inconsistency for this release, we have to revert this
change.

This reverts commit 6b896f43bef38c42d574d8bc89293766e0dbc94f.