Make Dir parameters const in all Stripe methods
I noticed while extracting methods and moving them to Stripe that I
could not mark the Dir parameter const. This is a followup to fix that
issue and apply const to all the Dir parameters in Stripe methods. This
makes the code more explicit and also helps the compiler generate
performant assembly.