Don't statically link libtsutil.a into core plugins.
Also, two equivalent libraries are now created, libtsutil.a and
libtsutil_link_dummy.so, with the same code and symbol content.  External
plugins can be linked with libtsutil_link_dummy.so  The drawback is, for
dlopen() to successfully load the plugin, it will require the presence of
libtsutil_link_dummy.so.  But, the symbols in libtsutil_link_dummy.so don't seem
to be used, because they are satisfied by symbols from core TS.  I tested this
(on Red Hat 8) by substituting a random .so for libtsutil_link_dummy.so, before
successfully loading the xdebug.so plugin.

I checked for an option in ld to resolve symbols for the link in a .so, without
creating a dependency on it.  But I could not find any such option.