Cleanup ink_thread.h of wrong BSD checks assuming missing pthread functions (#10863)
Commit 26c4080 attempted to replace some FreeBSD pthread function checks with function checking. This was reverted in 012df52 as feature detection should be done via autoconf (or now, CMake).

However, every remotely recent version of FreeBSD has the pthread functions in question. There seems to be no reason to bother to check for pthread_cancel or pthread_getschedparam.

In doing this, this allows us to remove HAVE_SYS_CAPABILITY_H and HAVE_SYS_PRCTL_H, replacing them with __has_include directives. We also introduce a direct HAVE_PRCTL feature check to replace the implied existence of it in HAVE_SYS_PRCTL_H.

In addition, the merge of #10823 into #11029 dropped a fix in which HAVE_PTHREAD_MUTEXATTR_SETTYPE was removed but not what called it. pthread_mutexattr_settype is available on every modern operating system supporting pthread; we should no longer need to check for its existence.