RegexContext instance created in the heap for each thread is leaked.
Damien found by running an ASAN build that the RegexContext instance created in the heap for each thread is leaked.
The thread_local RegexContextCleanup variables are never constructed becasue they are never referenced in any thread.

The solution is a different approach to avoiding deadlock when loading plugins with statically allocated DbgCtl
instances.  Construction of thread_locals is avoided while the DbgCtl registry mutex is locked.