Test failure checks for Stripe::add_writer (#11160)
This adds a new data-driven unit test to check all 32 conditions for the behavior of Stripe::add_writer and fixes an error in the docstring for that method.

There is no test to confirm whether or not the connection was added to the queue because that cannot be determined easily. I think that should be tested at the point where the write to the aggregation buffer occurs.

(cherry picked from commit 4c0bcfe9ca034ec602e373b471dcb7cf8cbf7a19)