Disable variable length array warnings
clang 18 warns about the use of dynamic variables used as the size of
arrays on the stack. We use this in a variety of places. I'm addressing
this by simply disabling the warning for now since we have used these
forever and my feel is that the community is OK with it. Some possible
alternatives would be to use alloca or to move this memory to the heap
via std::vector, but that, if desired, can be done in a separate patch.

Example warning:

src/api/InkAPI.cc:8421:16: error: variable length arrays in C++ are a Clang extension [-Werror,-Wvla-cxx-extension]
 8421 |       char buf[sid->len * 2 + 1];
      |                ^~~~~~~~~~~~~~~~
src/api/InkAPI.cc:8421:16: note: read of non-constexpr variable 'sid' is not allowed in a constant expression
src/api/InkAPI.cc:8420:27: note: declared here
 8420 |       const SSLSessionID *sid = reinterpret_cast<const SSLSessionID *>(session_id);
      |                           ^