std::shared_ptr::unique -> std::shared_ptr::use_count
std::shared_ptr::unique is deprecated since c++17 since it is
approximate in multithreaded environments. The inliner plugin made use
of unique. The one place that it used unique positively to reset data
it does so under a lock, so that seems safe. The other use of unique
was to compare !unique in such a way that seems flexible with the result
being approximate anyway. So from my inspection, it seems safe to simply
transition to use_count.