Fix HTTP/2 Buffering Issues (#11143)
When I ported #8747 into HTTP/2 to origin logic, I missed setting the watermark on outgoing connections. This essentially made it so that regardless of configuration, POST requests or the like would always send every little bit of DATA frames regardless of the configured watermark. The watermark was effectively zero on the server side. This patch fixes that by appying the configured watermark.

This patch also fixes the HTTP/2 write_time_threshold configuration so it consistently wakes up ATS to transmit the frames held up behidn write_buffer_threshold. If the write_buffer_threshold is not met, then I noticed ATS never woke up to send the accumulated frames until a xmit with a flush was sent at the end of the transaction. This ensures that the frames are transmitted per the write_time_threshold by scheduling an event for the retransmission.

Here's a list of changes made in this patch:

* Adding a write_threshold test
* Set buffer watermark for origin h2 connections
* HTTP2_SESSION_EVENT_XMIT -> HTTP2_SESSION_EVENT_PRIO
* schedule flushing event for write_time_threshold