Fix parent proxy connection stat overcounting (#11220)
The proxy.process.http.current_parent_proxy_connections and
proxy.process.http.total_parent_proxy_connections stats are each
incremented in create_server_txn, which is a fine place to increment
them, but this assumes that each server transaction is associated with a
single origin connection. This is not the case when sessions are reused.
This fixes the counting by only incrementing these stats for the very
first transaction in the session.