wasm plugin: Add build support (#11258)
* wasm plugin: Add build support

Initial commits of plugin/experimental/wasm on master had automake build
support. When we transitioned to cmake we temporarilly dropped the
automake files, leaving cmake support to a future commit. This is that
future commit.

* Silence -Werror=missing-field-initializers

This makes it so the lib/src/wamr/wamr.cc builds fine without a warning
concerning uninitialized members. Presumably the library writer knows
what they are doing.