Fix object and buffer lifetimes
There was a race condition where the callback handler got invoked
after it was freed, this corrects that by waiting for the callback.
This also frees the buffers allocated for the test, and makes a few
tweaks to the segment and bucket numbers to fix out-of-bound reads
caused by broken invariants among those numbers.