Remove reduntant link of ts:: libraries in plugins
This fixes the building of plugins that redundantly link with ATS core
libraries. If plugins link with ts::tsutil while ATS core also links
with ts::tsutil, for example, then not only is this unnecessary, but it
also results in ODR issues because those symbols will exist via both the
ATS and the plugin invocations.