Get Stripe::aggWrite into a test harness (#11178)
This demonstrates how to set up and perform a working write to disk
on a Stripe. The data is written to a tmpfile that the test holds
the file stream pointer to, so it should be possible to inspect and
verify the written data. The test does not do this, but having this
example may be useful.