Plugin, txn_box: Fix invalid use of signed index when the expected is (#11241)
unsigned.
It seems that the code was trying to deduce and store the kv separator
from
a query string, this does not make sense as the separator char is known
beforehand(used to split the kv), so instead, we just leave the default sep.

(cherry picked from commit 2a7aab10f55f1d1a71caf7a021bbdab256f3f50c)