Remove the need of `-Wno-stringop-overflow` warning suppression (#11357)
* Remove the need of -Wno-stringop-overflow warning suppression

The problem was that GCC complained for possibly incorrect usage of
strncat due to the last argument being equal to the source string length.
The last argument of strncat is supposed to be the remaining free space in
the destination buffer.
Replaced the usage of strncat with the ATS function ink_strlcat.

* Fix bogus warning for strncat usage in s3_auth plugin given by GCC 11

The change is to use TSstrlcat instead of strncat.
It seems to me the GCC analyzer is wrong in this case and it issues the
warning because of such usage:
str_concat(&left[loff], (left_size - loff), "/", 1);
and the str_concat internally does
std::min(dst_len, src_len) for the strncat size argument.
So GCC sees the possibility that the last argument of strncat can be
equal to the legnth of the source argument and issues the warning.

(cherry picked from commit 7220d700a20d333ceb9e051864d97ec98878c63f)