Test writing an evacuated document to the cache
This tests VCs with f.evacuator set to make sure all main code paths in
agg_copy are covered. It is very similar to the test for normal VCs, but
sets the buffer up with a document instead of the raw content string.
This commit also moves FakeVC and WaitingVC into a new file called
test_doubles.h.