Add tests for writing documents to the cache (#11347)
* Add tests for writing documents to the cache

This tests that the correct document data is written to cache by
aggWrite. It also covers some fields of the Doc struct used for reading
the data: the magic value, the length and header length, and the
optional checksum.

This commit also fixes the use of an uninitialized variable (header.phase)
reported by valgrind by initializing it to the same default value used
in the Stripe::_clear_init.

* Test writing an evacuated document to the cache

This tests VCs with f.evacuator set to make sure all main code paths in
agg_copy are covered. It is very similar to the test for normal VCs, but
sets the buffer up with a document instead of the raw content string.
This commit also moves FakeVC and WaitingVC into a new file called
test_doubles.h.