Add Remap ACL table test
This change adds a table test that tests every combination of ACLs,
namely:

1. remap.config inline rule
2. named ACLs in remap.config.
3. ip_allow.yaml.

It tests the expected response for a GET request (explicity denied or
allowed in the test) and a POST (implicitly denied or allowed by the GET
rule)

The current test implementation is optimised for minimal code changes.
Because of this a new server is started for each test. Because of this
the new tests alone will add 1m20s to the total autest testing time.

This could be optimised if we feel that this style of test is worth
considering.