Fix XpackDynamicTable::_make_space return value (#11408)
Callers expect `_make_space` to return true if it was able to free the
amount of given space, but it rather returns true if the amount of space
available exceeds the passed in value. This fixes the function to return
the intended result.

(cherry picked from commit b3b7a5674cd5c59c3470ec0597e43762ebaef60b)