Fix race condition causing interleaved logs
The output of the `DUMP_HEADER` macro was being written in multiple calls to
fprintf() that raced with other logging, resulting in interleaved output that
intermittently failed the ja3_fingerprint AuTest on CI. Whether other AuTest
failures were related to the same issue is unknown at this point.

This fixes the race condition by batching the calls to a single fprintf() call.
This also refactors the macro to two inline functions. In a benchmark using an
expanded version of the headers in the HttpTransact tests, a 60% speedup was
measured over the macro if all the header data fit in the buffer. In the worst
case that the buffer could only hold a byte at a time, the speedup was 80%.
The benchmark was compiled with -O3 and stderr was piped to /dev/null.

These were the headers used for the benchmark:
```
  header input1[] = {
    {"AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA", "111111111111111111111111"},
    {"BBB",                               "222"                     },
    {"CCC",                               "333"                     },
    {"DDD",                               "333"                     },
    {"EEE",                               "333"                     },
    {"FFF",                               "333"                     },
    {"GGG",                               "333"                     },
    {"HHH",                               "333"                     },
    {"III",                               "333"                     },
    {"JJJ",                               "333"                     },
    {"KKK",                               "333"                     },
  };
```

This fixes #11431.