Update magick plugin to OpenSSL 1.1.1 API
The functions `EVP_MD_CTX_create` and `EVP_MD_CTX_destroy` were renamed to
`EVP_MD_CTX_new` and `EVP_MD_CTX_free` in OpenSSL 1.1.0. This renames them
in the magick plugin accordingly. This also replaces the individual calls
to `EVP_DigestVerifyUpdate` and `EVP_DigestVerifyFinal` with a single call
to `EVP_DigestVerify` which is available in OpenSSL 1.1.1.